Merges branches

+++ LeMaitreCore/src/br/cefetmg/LeMaitre/model/dao/TableDAO.java
@@ -14,10 +14,10 @@ import java.util.List;
  * @author Thalesgsn
  */
 public interface TableDAO {
-    public Long insert(Table table) throws PersistenceException;
+    public Integer insert(Table table) throws PersistenceException;
     public boolean update(Table table) throws PersistenceException;
-    public boolean remove(Long tableID) throws PersistenceException;
-    public Table getTableByID(Long tableID) throws PersistenceException;
-    public boolean containsThisTableID(Long tableID) throws PersistenceException;
+    public boolean remove(Integer tableID) throws PersistenceException;
+    public Table getTableByID(Integer tableID) throws PersistenceException;
+    public boolean containsThisTableID(Integer tableID) throws PersistenceException;
     public List<Table> listAll() throws PersistenceException;
 }
